fix scripts and requires

+++ release/Magarena/scripts/Malignus.groovy
@@ -6,7 +6,7 @@
                 final MagicPlayer player,
                 final MagicPowerToughness pt) {
             // floor((x + 1) / 2) == ceil(x / 2)
-            final int amount = (player.getOpponent().getLife() + 1) / 2;
+            final int amount = (int)(player.getOpponent().getLife() + 1) / 2;
             pt.set(amount,amount);
         }
     },
